# backend/openapi.json
{
  "components": {
    "schemas": {
      "AttachmentResponse": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_image": {
            "title": "Is Image",
            "type": "boolean"
          },
          "original_name": {
            "title": "Original Name",
            "type": "string"
          },
          "post_id": {
            "title": "Post Id",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          }
        },
        "required": [
          "id",
          "post_id",
          "original_name",
          "content_type",
          "size",
          "is_image"
        ],
        "title": "AttachmentResponse",
        "type": "object"
      },
...
      "CommentCreateRequest": {
        "properties": {
          "content": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "CommentCreateRequest",
        "type": "object"
      },
      "CommentResponse": {
        "properties": {
          "author_id": {
            "title": "Author Id",
            "type": "integer"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "post_id": {
            "title": "Post Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "post_id",
          "author_id",
          "content",
          "created_at"
        ],
        "title": "CommentResponse",
        "type": "object"
      },
      "CommentUpdateRequest": {
        "properties": {
          "content": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "CommentUpdateRequest",
        "type": "object"
      },
      "ContentStatus": {
        "description": "A-03 / ADR-0005: only COMMITTED rows are served to users.",
        "enum": [
          "PENDING",
          "COMMITTED"
        ],
        "title": "ContentStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PostCreateRequest": {
        "properties": {
          "content": {
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "PostCreateRequest",
        "type": "object"
      },
      "PostDetailResponse": {
        "description": "Post GET response including committed attachments with thumbnail URLs (AC6 grid).",
        "properties": {
          "attachments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AttachmentResponse"
            },
            "title": "Attachments",
            "type": "array"
          },
          "author_id": {
            "title": "Author Id",
            "type": "integer"
          },
          "board_id": {
            "title": "Board Id",
            "type": "integer"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "board_id",
          "author_id",
          "title",
          "content",
          "status",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "PostDetailResponse",
        "type": "object"
      },
      "PostListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PostResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "PostListResponse",
        "type": "object"
      },
      "PostResponse": {
        "properties": {
          "author_id": {
            "title": "Author Id",
            "type": "integer"
          },
          "board_id": {
            "title": "Board Id",
            "type": "integer"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "board_id",
          "author_id",
          "title",
          "content",
          "status",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "PostResponse",
        "type": "object"
      },
      "PostUpdateRequest": {
        "properties": {
          "content": {
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "version": {
            "minimum": 0.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "title",
          "content",
          "version"
        ],
        "title": "PostUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Boards API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/boards/{board_id}/posts": {
      "get": {
        "operationId": "list_posts_boards__board_id__posts_get",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Posts",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "create_post_boards__board_id__posts_post",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Post",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/comments/{comment_id}": {
      "delete": {
        "operationId": "delete_comment_comments__comment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Comment",
        "tags": [
          "posts"
        ]
      },
      "put": {
        "operationId": "update_comment_comments__comment_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Comment",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/posts/{post_id}": {
      "delete": {
        "operationId": "delete_post_posts__post_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Post",
        "tags": [
          "posts"
        ]
      },
      "get": {
        "operationId": "get_post_posts__post_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Post",
        "tags": [
          "posts"
        ]
      },
      "put": {
        "operationId": "update_post_posts__post_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Post",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/posts/{post_id}/comments": {
      "get": {
        "operationId": "list_comments_posts__post_id__comments_get",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommentResponse"
                  },
                  "title": "Response List Comments Posts  Post Id  Comments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Comments",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "create_comment_posts__post_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Comment",
        "tags": [
          "posts"
        ]
      }
    }
  }
}